requirements.txt
certifi==2021.5.30
cycler==0.10.0
kiwisolver==1.3.2
llvmlite==0.37.0
matplotlib==3.4.3
numba==0.54.1
numpy==1.21.2
Pillow==8.3.2
pynndescent==0.5.4
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
umap-learn==0.5.1
wincertstore==0.2
